[0.68] Update autolinking to not error if a dependency is misconfigured (#10163)
While validating dependencies during autolinking, we threw a misleading error message in one very specific case of a module being misconfigured, but otherwise just logged the validation failure and move on.

This PR changes the behavior such that autolinking just logs the validation failure in all cases.

I recognize that we may *want* to throw errors and alert the user, but I think adding new errors is a breaking behavior change that users might not appreciate. I want this PR to be non-breaking and backportable (at least to 0.68), and I've opened #10013 to track how we might introduce that breaking behavior more thoughtfully.