Adds optional API for window sizing root
It may not always be the case that the XamlRoot is the same size as the
HWND. For example, in our app, to work around
https://github.com/microsoft/microsoft-ui-xaml/issues/2101, we resize
the XamlRoot content to be larger than the window by the DPI scale
factor and then resizes the ReactRootView to be the same size as the
HWND. However, this breaks the LogBoxModule and the AlertModule, which
make the assumption that the XamlRoot is the same size as the window.

To workaround this assumption, this change adds an API to XamlUIService
that allows you to specify which view under XamlRoot is equivalent to
the window size for the purpose of sizing things like the ContentDialog smoke
screen and the LogBox popup.

Fixes #10211