Check for DP::UnsetValue() rather than nullptr
When transferring properties in the FrameworkElementViewManager, we were
testing whether the result of `DP::ReadLocalValue` was null before
clearing the property on the old view and transferring the property to the
new view.

`ReadLocalValue` actually returns `DP::UnsetValue`, not nullptr, so this
check has been faulty and unnecessarily transfers unset values. This
commit switches the conditional to test for UnsetValue instead of null.