Use programmatic focus for UIManager.focus
Switching from xaml::FocusState::Programmatic to
xaml::FocusState::Keyboard causes focus rings to be shown for "default
focus" scenarios in our app which are not intended to have focus rings
(at least until the user explicitly changes focus with the Keyboard).

This change reverts to Programmatic focus in the UIManager.

Fixes #10265