Fix compiler issue for older versions of C++/WinRT
In older versions of cppwinrt, the header generated for
Windows.Foundation.Collections.h do not include the begin and end
methods for `IIterable`.

This just switches the logic for compatibility with older versions of
cppwinrt. Note, it may also include a bug fix where a comma separator is
not emitted between headers from the request header collection and
headers from the content header collection.