Use u8path for bundle loading (#10297)
* Use u8path for bundle loading

For side-loaded or unpackaged apps built with React Native, it's not a guarantee that the JS bundle will be in an ASCII path. For example, if the bundle is installed into the %LOCALAPPDATA% folder, it's possible that this path contains UTF-8 characters for the user name. `fs::path` does not support UTF-8 characters.

This change switches to u8path in case there are UTF-8 characters in the path where the bundle is located.

* Change files