Remove call to nyi() in TransformableText
We probably do not want this to crash the app in production, and it's
somewhat strange to depend on an inlined function from the
double-conversion.h that is not being used anywhere else. For now, just
switching to `assert(false)`, which should only crash in debug builds
and is used in other places throughout RN Windows.