Initialize TextTransformParentVisitor state
The textTransform logic is designed in a way that inherited transforms
work properly. To accomplish this, when any change is made to raw text,
it traverses its parent tree to find the first inherited transform.

When the raw text node is first created, it does not find any parent, so
it uses the default initialized state value for textTransform in the
TextTransformParentVisitor. Depending on compiler settings, this is not
always 0 / undefined.

This change ensures the textTransform parent state is always set to
Undefined.