Fix crash in DynamicAutomationPeer::GetContentName (#10321)
* Fix crash in DynamicAutomationPeer::GetContentName

We occasionally see a crash when the ViewPanel is null while running
this method. While we don't have a consistent repro, this change
resolved the issue. My guess is that there is some edge case where the
ViewControl can be unmounted, calling `removeAllChildren`, which cleans
up the `ViewPanel`, but somehow the automation peer still holds a
reference to the ViewControl and tries to do work.

Co-authored-by: Liron Yahdav <lyahdav@users.noreply.github.com>

* Change files

Co-authored-by: Liron Yahdav <lyahdav@users.noreply.github.com>