Make borderRadius consistent with other RN platforms
The `borderRadius` prop on other platforms (including Web) is clamped to
the minimum dimension between width and height. On Windows, without this
change, borderRadius is delegated to the `CornerRadius` XAML property,
which will produce an oval shape from the `borderRadius` rather than
clamping.

This change is a proposal to imperatively clamp the borderRadius, though
there are likely more efficient ways of accomplishing this.

Resolves #10355