Restore Git Tags for Canaries (#10256)
## Description

### Type of Change
- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)

### Why
Years ago the RNW repo published new package revisions for every commit to the repo. Beachball's default is to create a new git tag for every package revision. This led to a Git tag, for every package, for every commit. GitHub previously included every tag  as an indivdual release in its "Releases" view. This drowned out the public-facing release notes, leading to a change to only tag package revisions where release notes were wanted.

GitHub has changed their UI, to no longer include tags in the "Releases" view. A current example where there are many tags but no releases is https://github.com/microsoft/fluentui-react-native/releases. This decoupling allows walking back sparseness of tagging.

### What
This change adds git tags for canaries and for the CLI package. These can be useful for local debugging, and provide a way to correlate package versions to a git revision, without having to embed a commit hash into the version. Tags created by beachball conform to the standard of `{package}_v${version}`, including for scoped and prerelease packages.

This requires changing a bit of logic in `@rnw-scripts/create-github-releases` as well, which currently assumes it should automatically create releases for anything with a Git Tag.