Only fire `onLongPress` for left clicks (#10325)
* Only fire `onLongPress` for left clicks

In #9603, we re-introduced right click events to react-native-windows.
We specifically prevent `onPressIn`, `onPressOut`, and `onPress` events
for anything other than the default button (e.g., left click). We should
do the same for right click events. Note, it is still possible to
emulate "right click" behaviors for long press on touch devices as these
types of presses are still treated as the default button.

This change checks if the event uses the default button before firing
the onLongPress event.

* Change files

* yarn lint

Co-authored-by: TatianaKapos <tatianakapos@microsoft.com>