Adds Systrace to view management APIs for UIManager
This change adds SystraceSections to many relevant UIManager and View
calls for more visibility into what work is being done by RN.

Specifically, it instruments the following:
1. Most of the relevant view management APIs for the top-level UIManager
and the NativeUIManager (e.g., createView, updateView, etc.). It does
not currently measure any context switching from JS to UI thread.
2. Potential hot spots for ViewViewManager, like replacing the ViewPanel
with a ViewControl and explicit time spent in MeasureOverride and
ArrangeOverride for XAML layout.
3. Time spent preparing for and executing Yoga layout, including
extra XAML layout calls made to prepare the view for Yoga layout (e.g.,
UpdateExtraLayout and NeedsForceLayout), the call to
YGNodeCalculateLayout, and the calls to SetLayoutProps.