Fixes various issues with fast text optimization
In #10354, a few scenarios are described where unexpected text changes
occur out of sync with the shadow tree from RN, and some scenarios even
crash the app.

This change ensures that we only use the fast text optimization when
there is exactly one raw text child in the shadow tree, and also
switches out of the fast text optimization when the number of children
increases to greater than 1, rather than only when a text value is
appended (i.e., we also have to switch out of fast text optimization
when for a child that might be prepended at index 0).

Fixes #10354