Enable ABIViewManager to opt out of NeedsForceLayout
Currently, setting RequiresNativeLayout to true also opts into
NeedsForceLayout on the shadow node, which makes an extra UpdateLayout
call in UIManager. These two features are not necessarily tied together.
For example, TextInputViewManager sets both RequiresNativeLayout and
NeedsForceLayout, but TextViewManager does not.

We are currently experiencing a crash when being forced to call the
extra UpdateLayout call on a custom view manager based on
ABIViewManager.

This change adds another interface that allows ABI VMs to opt out of the
forced extra UpdateLayout.