Adds VM interface for custom measurement
While the default self measure function implemented in ViewManagerBase
works in most scenarios, it may not work for all use cases. This change
introduces an ABI VM interface to callback when Yoga measurement is
being performed.

Since Yoga self measure functions are C-style function pointers, we
can't pass instance methods to Yoga. Thus we needed to use a trampoline
approach to provide a static method that can subsequently call the
appropriate measure instance method on the view manager (that has access
to the IViewManagerWithNativeLayout intsance as well). In order to
support this, we needed to add a pointer to the view manager to the
YogaContext that gets passed back to the measure function.

Adding the VM pointer to YogaContext will add 8 bytes to each Yoga node, which is
not great, but YogaContext is only added to Text, TextInput, and any ABI
view managers with native layout, so this probably shouldn't be a
concern.