Adds gesture cancellation to pointer event handling (#10203)
* Adds gesture cancellation to pointer event handling

In https://github.com/microsoft/react-native-windows/pull/8495, we added
an interface for 3rd party view managers to override pointer event
handling in the TouchEventHandler module from RNW. The features of this
PR supported press events on selectable text, but to generalize this
further, we need to allow 3rd party view managers to conditionally
cancel pointer event sequences and prevent press events.

This change adds the ability for view managers implementing the
IViewManagerWithPointerEvents interface to change an "End" event to a
"Cancel" event to achieve just that.

Fixes #10202

* Change files