Fix XamlRoot crashes in FlyoutViewManager (#10314)
* Fix XamlRoot crashes in FlyoutViewManager

There are two cases where a flyout may crash when opened:

1. If the native view corresponding to the target tag is unmounted before the flyout is shown.
2. If the XamlRoot on the flyout differs from the XamlRoot of the target.

This can happen if the XamlRoot changes (e.g., window loses focus) before the UIManager operation to create the Flyout and set the isOpen prop executes. This change logs the error and does not attempt to show the flyout if either of these cases occur.

* Change files