Ensures pointer event is updated when leaving bounds of React root
Previously, if we did not find a valid React target, we would not update
the pointer data, resulting in firing the `onTouchEnd` with the last
data used while the pointer was still inside the React root.

This change ensures that, while the pointer is down (i.e., we have a
captured pointer in the React root), we continue to update the pointer
data and also emit move events. The viewX / viewY values of the pointer
will be relative to the declared root view.

Fixes #10406