Add `$schema` to `cgmanifest.json` codegen (#10495)
## Description

This PR correctly implements the fix attempted in #10474, by updating the code-gen of our `cgmanifest.json` files, rather than the output.

### Type of Change
- Bug fix (non-breaking change which fixes an issue)

### Why
The change to add a `$schema` field to the `cgmanifest.json` files, while righteous, was applied to generated files, and to only one of those present in our repo. This means the changes will be lost during the next build.

### What

Updated and re-ran the code-gen to add the `$schema` field.