Fixes bug introduced by pointerEvents change
The handling for pointerEvents moved from ViewViewManager to ViewManagerBase so all views could participate in pointerEvents handling (including Text, etc.). During this move, the code introduced a bug where updates to pointerEvents after pointerEvents is set to `none` are not respected.

This adds an else clause to ensure we reset the `IsHitTestVisible` property when a the `pointerEvents` value is not `none`, but still valid.

Resolves #10493