Fixes event ordering for `onLayout` (#10422)
* Fixes event ordering for `onLayout`

Other platforms fire `onLayout` depth first. Windows should do the same.
Also, once we upgrade to React 18, it is a common case that React
commits will create nodes that are not yet attached to any root (and
thus do not participate in the Yoga layout for that batch / commit).
This means that, in the current approach for updating layout, we would
fire `onLayout` events with `NaN` values for width and height.

This changes switches from iterating over all Yoga nodes to update
layout on each UIManager batch to recursively updating layout from the
root shadow node. The effect is that we fire events depth first and we
only attempt to update layout on nodes that have participated in Yoga
layout (because they are descendants of some root).

This also likely fixes a bug in our app we were seeing where attempt to
transfer an explicit NaN Width/Height property when replacing a
ViewPanel with a ViewControl or Border (for relevant prop
implementations) would occasionally throw an exception.

Fixes #10421

* Change files

* Use GetYogaNode instead of lookup in m_tagsToYogaNode

* Move SystraceSection header to root set layout call

* Avoid unnecessary recursion for nodes that do not have Yoga children

* yarn format