Remove unnecessary tree walk in Yoga layout (#10497)
* Remove unnecessary tree walk in Yoga layout

I believe this code is dead code that is no longer needed. It was originally introduced to apply Yoga paddings to ContentControlViewManager from the Polyester folder. However, since then, the approach to handling controls that implement their own padding has changed.

The code that was removed effectively walks all nodes attached to a root node every time the root view size changes or a batch of UI operations completes. Not anymore :).

* Change files