Adds event to listen for quirk settings changes
If we want to set QuirkSettings at runtime, and some component would
like to subscribe to those changes, there is currently no mechanism to
do that. This adds a winrt::event and a few internal hooks to subscribe
to QuirkSettings changes.

The plan is to use this in the PaperUIManager module for experimentation
with a few quirk settings, like view flattening or setting
`removeChildren` to false when calling `DropView` from `manageChildren`.