Adds quirk setting to skip removing children in unmount
The `removeChildren` option in `DropView` can be used to conditionally
recurse the Yoga and native view tree to clean up each node. This would
be required if we wanted to pool or otherwise recycle Yoga nodes or XAML
views, but we do not do this today.

While we could just explicitly set this to false, my concern is that
this could create potential memory leaks, and its better to default to
the pre-existing behavior and opt-in for experimentation.

In addition to adding the QuirkSetting, this change also wires up a
change event on QuirkSettings that can be subscribed to in
PaperUIManagerModule so we can get fresh quirk settings when they
change. While quirk settings are generally set before starting the
ReactInstance, if someone wanted to A/B test this functionality and hide
it behind a feature flag, it's possible that they might want to change
the QuirkSetting at runtime.