Adds view flattening to Paper UIManager
Most react-native apps are littered with views used only for
positioning / layout. Creating these controls is a costly in terms of
CPU time spent constructing them, adding them to the native XAML tree
and time spent running the measure / arrange pass in XAML.

This change implements view flattening to avoid creating native backing
views for layout-only nodes.

The core of the implementation is in the set of functions declared in
ViewFlattening.h. It abstracts away methods that were previously invoked
directly on shadow nodes, and instead performs tree-traversals to find
the correct location in the native tree that a given node should be
attached or detached.

The rest of the changes are auxiliary:
- ShadowNodeBase adds a virtual method to report whether the node is
layout only.
- ShadowNodeBase adds a 64-bit integer to track the number of non-layout
only children it manages to avoid recursion into each node any time a
view is added, removed, or otherwise re-arranged.
- ViewManagerBase adds the `collapsable` native prop type, which is
more-or-less an override flag for forcing materialization of a
layout-only view.
- ViewViewManager is updated to avoid creation of a native view if the
initial props are layout-only. It also checks the layout-only status
when props are updated so it can create the native view if and when it's
needed.
- PaperUIManager is where previous native view hierarchy operations were
performed directly via ShadowNodes, but are now performed via
ViewFlattening functions
- NativeUIManager adds functionality to support setting layout on
children of layout only nodes, as well as functionality to re-apply
layout when a layout-only node transitions to a materialized view. There
are also changes to measurement APIs to accommodate for non-materialized
views.