Avoid batching create view operations
iOS and Android each have their own approach to create view components
immediately when the `UIManager.createView` operation is invoked. On
Android, the `createView` operations are aggregated in a separate batch
and allowed to run for half a frame. On iOS, the `createView` operations
always immediately create the native view when the UIManager module
method is called.

On Windows, we have been adding the `createView` operation to the batch
of operations that invoke when the batch is completed, missing an
opportunity to better utilize the UI thread while the JS thread is still
busy doing work.

We ran an experiment for this on Messenger with some significant
results. Time-to-rendered-content (TTRC) for important scenarios like thread
switching improved by more than 8%, cold start latencies improved by
~1%, overall memory usage improved by a small margin, and TTRC for
message sends also improved by about 3%.

Please note, this should not create any UI tearing, as the operations to
mount the created views in the native XAML tree are still only performed
on batch completion.