Fixes issues with packaged asset resolution
`Image.resolveAssetSource` has two different code paths - one for
resolving assets from the packager server when running in standard debug
builds and one for resolving assets when the JS package is pre-bundled.

Previously, we were not providing the bundle root path to the
`SourceCodeModule` for pre-bundled apps, and it was up to each
individual native module and view manager that depends on resolving
asset paths with `Image.resolveAssetSource` to replace `file://` with
the bundle root path pulled from the InstanceSettingsSnapshot.

Even after fixing the issue with `SourceCodeModule`, there is still one
more issue for unpackaged apps, whose asset paths may be Windows file
paths (e.g., `C:\...`) rather than packaged asset paths (e.g.,
`ms-appx://`).

This change ensures that SourceCodeModule provides the bundle root path
as the `scriptURL` constant and patches the bug in the upstream
`resolveAssetSource` module to skip coersion of the bundle root path to
a form like `file://`.

Resolves #10620