Suppress Flyout if XamlRoot does not have focus
If attempting to show a flyout on an unfocused window, the flyout will
show on top of the window currently in focus and appear unattached to
anything in particular.

This change ensures that we only attempt to show a flyout if the
XamlRoot attached to the flyout is the same as the active XamlRoot set
in XamlUIService.