Prevent crash in XAML Islands Flyout
A recent change to FlyoutViewManager attempts to set the XamlRoot from
the NativeUIManager::tryGetXamlRoot function. However, there is a race
condition here where the view could be created just after the window was
closed (or perhaps even that the window was closed just after the view
was created and the weak reference to the XamlRoot on the Flyout
resolves to null). When this occurs, attempting to call Flyout::ShowAt
results in a crash in XAML Islands applications.

This change ensures that the Flyout has a valid XamlRoot before
attempting to show the flyout.