Refactor the onLayout event for view flattening
We are working on an implementation of view flattening for the Paper
UIManager in react-native-windows. Since not all nodes are materialized
into native components, the backing store for previous layout data
cannot be the native component itself. Also, the actual top/left values
that get applied to a particular view will not equal the values computed
by Yoga (they will be offset by the cumulative left and top values of
any flattening parents).

This change adds 4 bytes to the ShadowNode to store the latest layout
constraints computed by Yoga and lifts the `onLayout` event to the
NativeUIManager, so that even if a view is flattened, it can still emit
`onLayout` events.