Adds QuirkSetting to use YGExperimentalFeatureWebFlexBasis
By default, Yoga will cache flex basis values to avoid expensive
recomputation. However, there are some scenarios, particularly when
resizing the window, where cached layout values from Yoga combined with
cached flex basis values can produce incorrect results.

This change adds a new QuirkSetting to opt-in to the
YGExperimentalFeatureWebFlexBasis option in Yoga for apps where re-use
of this cached flex basis value is problematic. We'll keep this opt-in
for now, as it is likely to impact performance of Yoga layout.

Resolves #9343