Adds LayoutService interface to allow native layout updates
In order to enable native layout updates, we need to expose an interface
to mark Yoga nodes as dirty, and also to trigger native layout.

While we could just blindly invoke `onBatchComplete` on the
NativeUIManager, it is safer to also expose an interface to test whether
the UIManager is currently processing a batch of UIManager operations
so we can piggy back on the current onBatchComplete call that will
occur at the end of a batch if one is in progress. This is also likely
to prevent UI tearing. Today, it is unlikely that an call from the UI
thread to invoke Yoga and XAML layout will interrupt a batch, if we
eventually implement something like eager view creation, this may no
longer be the case.

Also, rather than calling `onBatchComplete`, which may have the
side-effect of calling UpdateLayout on nodes that specify
`NeedsForceLayout`, native UI updates will not involve the creation of
new nodes, so we can skip this step, thus the addition of a new API to
call YGNodeCalculateLayout for a specific root (and skip layout for
other roots where it may not be needed). The LayoutService can also
conditionally apply layout to all roots, a specific root with unbound
constraints, or a specific root with known constraints.

We will be using this service for a number of features, including
re-entrant Yoga layout for views that are self-measuring with React
Native descendants dependent on Yoga layout, as well as to work around a
text truncation bug caused by Yoga caching to force the Text to
re-measure and re-layout when truncation is detected (is narrow use
cases).

Resolves #9988
Towards #7601
Workaround for #9343