Workaround for blur on XAML Islands
In XAML Islands, when a TextBox has focus, FocusManager will return
null, causing it to fail the check here:

https://github.com/microsoft/react-native-windows/blob/dbe1ac75d8df2e749782999e09b70023c218a7a4/vnext/Microsoft.ReactNative/Modules/NativeUIManager.cpp#L1117

This change checks if the view is a XAML Control, and if it is, checks
the FocusState. This is a sufficient workaround for XAML Islands.
However, this workaround is insufficient for custom controls like
the Hyperlink inline (which react-native-windows doesn't really support
well anyway).

Resolves #10690