Explicitly measure relative to the root XAML element (#10665)
* Explicitly measure relative to the root XAML element

Previously, `measureInWindow` measured relative to
`xaml::Window::Current().Content()`, which is null for XAML Islands
applications. Calling `TransformToVisual` with a null argument measures
relative to the root XAML content, which is equivalent to
`xaml::Window::Current().Content()` for the foreground window. The main
difference is that this API will now measure correctly for React Native
windows that are not in the foreground.

* Change files