Moves focus check to ShadowNode so arbitrary views can participate
The previous approach only fixed the issue for Controls, but if an app
has a 3rd party control type that sets the focused component on
TextInputState, they would not be able to participate in this. We can
eventually also include an ABIViewManager API to override the
`IsFocused` method, but that's not needed for now, and the behavior will
still be identical on UWP given we still check
FocusManager::GetFocusedElement.