Skip setting XAML layout when tree is collapsed
When a view tree is collapsed (via `display: 'none'` style setting),
there is no need to update the ViewPanel positioning properties
(Canvas::TopProperty, Canvas::LeftProperty,
FrameworkElement::WidthProperty, and FrameworkElement::HeightProperty),
which would in turn trigger invalidate XAML layout for ViewPanels in
that tree.

This change ensures we do not update these XAML positioning properties
when a view is in the context of a collapsed sub-tree.