Reduce frequency of arrange invalidation for ViewPanel
ViewPanel top and left props are set from Yoga results. For better or
worse, float values in Yoga are treated as equal if the difference is
less than 0.0001. As such, it's safe for us to only invalidate the
parent when the values are not equal with an epsilon of 0.0001.

We spend a non-trivial amount of time in ArrangeOverride for ViewPanel.
The Top and Left properties are set any time a Yoga node has a new
layout, which is not always when the actual layout value has changed
(e.g., only the width or height may have changed, or perhaps a
descendant value layout has updated, which often marks the parent as
having a new layout when nothing has actually changed).

This change should allow us to only invalidate the XAML parent when Yoga
layout has actually changed.