Mitigate MSB4011 warning for importing ExperimentalFeatures.props
This PR fixes the `MSB4011` warning when `JSEngine.props` tries to
import `ExperimentalFeatures.props` a second time. The expectation is
 that all RNW projects (app or module) import the
`ExperimentalFeatures.props` file (located next to the VS solution)
 at the very top of each project file, so all RNW-related projects
are in sync with resepect to the state of any Experimental Features.

However, older RNW apps/module templates did not include this import,
hence `JSEngine.props`'s workaround to import it. Unfortunately,
MSBuild:

* Doesn't like project files to be imported twice, so it gives the
  `MSB4011` warning
* Doesn't provide any built-in way to detect if a project file has
  already been imported
* Doesn't allow you to suppress this particular warning completely

The best solution is to set a property in the imported file and check
for that property before attempting to import the file.

This PR adds an `ExperimentalFeaturesSet` property to all in-repo and
new template `ExperimentalFeatures.props` files, and changes
`JSEngine.props` to both check for it before importing and throw a new
warning so let existing projects know about the new property and how to
resolve all of the warnings.

TLDR: After this PR has been merged, adding

```xml
<ExperimentalFeaturesSet>true</ExperimentalFeaturesSet>
```

to your `ExperimentalFeatures.props` file will get rid these annoying
MSB4011 warnings.