Fixes issue for loading images in Clang builds (#10772)
* Fixes issue for loading images in Clang builds

We recently attempted to upgrade cppwinrt in our app, and noticed that
image load failures were occurring. It seems that our combination of
cppwinrt and Clang doesn't like when you immediately co_return a
nullptr in a coroutine.

To work around this, this change ensures we only call into the coroutine
method if necessary (by first checking the `fromStream` boolean).

* Change files