Update WindowsAppSDK versions to latest (#10771)
Description

This PR updates the following versions for the `Microsoft.ReactNative.WindowsAppSDK` project and the template for apps built against it:

* Upgrade `WindowsTargetPlatformVersion` / `Microsoft.WindowsSDK` to 10.0.19041 from 10.0.18362
* Upgrade `WinUI3` / `Microsoft.WindowsAppSDK` to 1.1.4 from 1.0.0

Type of Change
- Breaking change (fix or feature that would cause existing functionality to not work as expected)
- This change requires a documentation update

Why

Building the `Microsoft.ReactNative.WindowsAppSDK` project with the `WindowsTargetPlatformVersion` at 10.0.18362 means all of the dependencies (like `Microsoft.ReactNative`) also build targeting that version, which causes multiple warnings by the MIDL compiler. Updating to 10.0.19041 aligns WindowsAppSDK with the standard UWP projects.

To make that upgrade/alignment happen, the `Microsoft.WindowsAppSDK` also needed to be bumped.

What
Updated versions in project files, including the WinAppSDK app template.