Apply fast text optimization to multiple runs
Currently, the fast text optimization is only applied for Text nodes
with a single run child. If you have text like the following:
```
<Text>Count: {count}</Text>
```
This will create two inline children.

We can instead validate that the Text node has only runs for children,
and concatenate the context of the runs together to keep using the
optimized text path (setting the xaml::TextBlock::TextProperty).