Fixes issue with zoom and horizontal scrolling (#10807)
* Fixes issue with zoom and horizontal scrolling

In #7072, a change was made to always disable horizontal scrolling by
default. Prior to this change, apps that enabled vertical scrolling and
zoom scale (via maximumZoomScale or minimumZoomScale) could scroll in
both directions. This change reverts the problematic commit from #7072
and puts in a better fix for the scrollToEnd issue.

* Change files