Workaround for blur on XAML Islands (#10691)
* Workaround for blur on XAML Islands

In XAML Islands, when a TextBox has focus, FocusManager will return
null, causing it to fail the check here:

https://github.com/microsoft/react-native-windows/blob/dbe1ac75d8df2e749782999e09b70023c218a7a4/vnext/Microsoft.ReactNative/Modules/NativeUIManager.cpp#L1117

This change checks if the view is a XAML Control, and if it is, checks
the FocusState. This is a sufficient workaround for XAML Islands.
However, this workaround is insufficient for custom controls like
the Hyperlink inline (which react-native-windows doesn't really support
well anyway).

Resolves #10690

* Change files

* Adds blur example to TextInput sample

* Moves focus check to ShadowNode so arbitrary views can participate

The previous approach only fixed the issue for Controls, but if an app
has a 3rd party control type that sets the focused component on
TextInputState, they would not be able to participate in this. We can
eventually also include an ABIViewManager API to override the
`IsFocused` method, but that's not needed for now, and the behavior will
still be identical on UWP given we still check
FocusManager::GetFocusedElement.

* Switch to GetFocusedElement(XamlRoot) overload