Use standard Panel APIs for children in ViewViewManager (#10890)
In preparation for a change that moves away from a custom ViewPanel
implementation to core XAML controls (like Grid and Canvas), this change
removes the dependency on ViewPanel wrapper calls to manage children.

It also changes ViewControl to return a generic XAML Panel type rather
than ViewPanel. There are no callsites in react-native-windows where
this must be a ViewPanel type, but since this is API is exposed via the
ViewControl ABI to third parties, this must be considered a breaking
change.