Update rnw-image* config backups and stop extra .NET installs
This PR reflects changes to the ADO VM images to update some of the
software we need to the latest (minor) versions.

These were updated after discovering that some of the Guardian tasks
were modifying the machines with new .NET SDK versions in a way which
breaks our builds.

This PR also updates the shared variables to prevent Guardian from doing
this again.

Closes #10951