Adds logic to retain undo/redo history for TextInput (#10133)
* Adds logic to retain undo/redo history for TextInput

As described in #10132, controlled text updates from JS to XAML TextBox
for TextInput cause the undo redo history to be lost because
imperatively setting the XAML TextBox::Text property resets history on
the TextBox. This change finds a single diff in the before and after
string and uses the SelectedText property which retains undo/redo history.

Fixes #10132

* Change files

* yarn format

* Update vnext/Microsoft.ReactNative/Views/TextInputViewManager.cpp

Co-authored-by: Chris Glein <26607885+chrisglein@users.noreply.github.com>

* Update vnext/Microsoft.ReactNative/Views/TextInputViewManager.cpp

Co-authored-by: Chris Glein <26607885+chrisglein@users.noreply.github.com>

Co-authored-by: Chiara Mooney <34109996+chiaramooney@users.noreply.github.com>
Co-authored-by: Chris Glein <26607885+chrisglein@users.noreply.github.com>