Fix raw text update clearing Text
In #10811, a bug was introduced where a raw text update to a
non-optimized Text component would call ClearValue on the
TextBlock::TextProperty. This behavior was introduced to ensure that we
cleared out the TextProperty before transitioning from optimized to
non-optimized text.

To fix this bug, we check that the text is actually optimized before
clearing the TextProperty, and for the case of transitioning text from
optimized to non-optimized, added an optional flag that must be set to
true before attempting to clear the text.

Please note, we never transition from non-optimized to optimized text,
so we don't need to worry about that case.