Mark Text dirty when descendant Text updates occur
Text should be marked dirty when descendant text nodes are added or
removed, and Text should also be marked dirty when descendants of
descendant virtual text nodes have changes to their text content.

This change ensures that the Yoga node for Text is appropriately updated
when these types of changes occur.

Please note, this is likely to create a slight performance regression
for apps that modify nested text content after initial mount, as these
changes previously would not trigger text measurement, and now they
will. But it's worthwhile to trade correctness for a minor performance
benefit.

Resolves #11007