Switch ViewPanel base class to Grid
Grid supports border properties (BorderThickness, BorderBrush,
CornerRadius) intrinsically.

We experimented with using Grid as the base class for ViewPanel in
Messenger and saw no regression in memory usage, and a moderate
improvement in time-to-rendered-content for important scenarios (as much
as 3% improvement in some cases).

In addition to the performance improvement, avoiding the complicated
outer / inner border logic in ViewPanel allows us to remove over 250
lines of code (more if we could make breaking changes to the ViewPanel
ABI) in ViewViewManager and ViewPanel.

Lastly, using Grid as the base class for ViewPanel allows third-party
view managers to create View-like custom native components that also
support border properties, without needing to rewrite the complex
inner/outer border management logic.