Fix search of react-native module code in vscode (#11088)
By default, vscode uses the `.gitignore` file to filter out files when
searching and for intellisense. This is usually good because of how many
dependencies RNW has that we don't ever look at.

However this limits our ability to easily browse, search, and jump into
code we *do* actually want to look at, especially all of the shared C++
in RN (which is only increasing in the new architecture).

This change updates our repo's settings file and adds a new `.ignore`
file where we can selectively allow certain folders under
`node_modules` to improve the dev experience in vscode.