Remove invalid accessibilityValue from SwitchExample (#11097)
The `accessibilityValue` prop is supposed to be an object with 4
properties: 'now', 'min', 'max', 'text'. The SwitchExample override for
RNTester on Windows sets the accessibilityValue to a boolean value,
which is not correct. For now, this just removes the accessibilityValue
in this example as it will be problematic for prop deserialization in
Fabric.