Fixes issue where Alert crashes
There are edge cases where an Alert may be shown while the app is
exiting. Generally, when the app is exiting, apps will have set the
XamlRoot instance on XamlUIService to null. In this case, we may still
attempt to show the alert on a XAML Island with a null XamlRoot
instance, which would cause a crash.

This change just ensures that, if we are in a XAML Island and we don't
have a valid XamlRoot to attach to the ContentDialog, we just treat the
alert as dismissed.