Ensure NativeAnimatedNodeManager is destroyed on UI thread
Most operations in the NativeAnimatedNodeManager run on the UI thread,
including operations that attempt to resolve Paper shadow nodes, and
operations to run the tick-driven animations. It's possible that the UI
thread operations may run after the NativeAnimatedNodeManager has been
destroyed, in which case, we see crashes when unloading the React
instance because we're referencing fields on `this` that no longer
correspond to a valid NativeAnimatedNodeManager instance.

This change uses the same pattern we use in the PaperUIManagerModule to
ensure that the UIManager instance is cleaned up on the UI thread, by
posting the shared_ptr to the instance to the UI thread if it's available.